requirements.txt
matplotlib
numpy
pandas
pytest
tensorflow-gpu